everything done
Former-commit-id: 1dcc1e94e6b95859845d937062ccafe94ece4675
Former-commit-id: a65c2126d7a0cb7f511f1bade96e40a54c9b9475

+++ QZBQuizBattle/QZBEndGamePointsCell.m
@@ -51,12 +51,12 @@
 
     UILabel *label = (UILabel *)self.circleView.centralView;
     label.textColor = color;
-    label.text = [NSString stringWithFormat:@"x%ld", multiplier];
+    label.text = [NSString stringWithFormat:@"x%ld", (long)multiplier];
 }
 
 - (void)setScore:(NSUInteger)score {
-    self.pointsLabel.text = [NSString stringWithFormat:@"%ld", score];
-    self.pointsNameLabel.text = [NSString endOfWordFromNumber:score];  // REDO
+    self.pointsLabel.text = [NSString stringWithFormat:@"%ld", (unsigned long)score];
+    self.pointsNameLabel.text = [NSString endOfWordFromNumber:score]; 
 }
 
 //- (NSString *)endOfWordFromNumber:(NSInteger)number {
